+++ test-demo.js
@@ -1,2 +1,0 @@
-// TODO: refactor this mess later
-console.log('WIP commit');